pyproject.toml
[tool.poetry]
name = "talamus-relatable"
version = "1.0.2"
description = "A relational database -like table of rows."
authors = ["Tero Niemi <talamus@gmail.com>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/talamus/python-relatable"
packages = [{include = "relatable"}]
classifiers = [
    "Topic :: Software Development :: Libraries :: Python Modules"
]

[tool.black]
exclude = "\\.venv|\\.git|\\.vscode|test_basics\\.py"

[tool.poetry.dependencies]
python = "^3.10"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.2.0"
pyyaml = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
